requirements.txt
flask
flask-sqlalchemy
flask-migrate
flask-marshmallow
marshmallow-sqlalchemy
flask-bcrypt
flask-jwt-extended
python-dotenv
mysqlclient
stripe
flask-cors
pyjwt
